requirements.txt
python-barcode==0.14.0
streamlit==1.23.1